refactor(baseline): split path and body-hash units out
`baseline.rs` sat at loc.ploc 524 against this repo's 550 file limit —
inside the 0.95 soft band `make self-scan-headroom` gates — and mixed
three unrelated concerns that the file was already ordered by.

Move them out verbatim into a module directory:

  baseline/path_key.rs   anchor resolution, lexical `.`/`..` folding,
                         and the percent-encoding pipeline that turns
                         an OsStr into a TOML-safe identity key
  baseline/body_hash.rs  the normalised-body FNV digest behind
                         `--baseline-fuzzy-match`, its hex codec, and
                         `bare_name`

`baseline.rs` keeps the file format, matching, and rendering, and
re-exports `anchor_for` / `hash_body` / `bare_name` so no caller
outside the module changes. Each module takes the tests that cover it,
in its own `*_tests.rs` sidecar; only `use` paths and visibility moved.

No behaviour change: baselines written before and after the split are
byte-identical, a pre-split baseline reads back to the same
classification, and `bca diff-baseline` / `bca exemptions` output is
unchanged. The full offender set is identical apart from the dropped
`<file>` loc.ploc offender and an 8-line shift in two spans.

loc.ploc: baseline.rs 524 -> 366, path_key.rs 88, body_hash.rs 81.
